feat: separate running commands + rewrite handling to websocketstream
