+++ src/websocket-server/index.ts
@@ -1,8 +1,5 @@
-import { WebSocketServer } from "ws";
-import { drawHandler } from "../draw/draw";
-import { parseCommand } from "../helpers/helpers";
-import { mouseHandler } from "../mouse/mouse";
-import { psHandler } from "../ps/printScreen";
+import { createWebSocketStream, WebSocketServer } from "ws";
+import { runCommand } from "../cmd/commands";
 
 export const startWsServer = ( port: number ) => {
   const wsServer = new WebSocketServer({ port: port });
@@ -13,35 +10,42 @@ export const startWsServer = ( port: number ) => {
 
   wsServer.on('connection', (ws) => {
     console.log(`WebSocket Server: Client connected`);
-    ws.on('message', async (data) => {
-      console.log('WebSocket Server <- %s', data);
 
-      const [command, arg1, arg2] = parseCommand(data.toString());
+    const duplex = createWebSocketStream(ws, { encoding: 'utf8', decodeStrings: false });
 
-      if (command === 'mouse_up' || command === 'mouse_down' || command === 'mouse_left' || command === 'mouse_right') {
-        mouseHandler(command, arg1);
-        ws.send(`${data}`);
+    duplex.on('readable', async () => {
+      let data;
+      while (null !== (data = duplex.read())) {
+        console.log('WebSocket Server <- %s', data);
+        const result = await runCommand(data);
+        if (result) {
+          if (result.startsWith('prnt_scrn')) {
+            console.log('WebSocket Server -> prnt_scrn <base64 string (png buf)>');
+          } else {
+            console.log('WebSocket Server -> %s', result);
+          }
+          duplex.write(result);
 
-      } else if (command === 'mouse_position') {
-        const posStr = (await mouseHandler(command)) as string;
-        ws.send(posStr);
-        console.log('WebSocket Server -> %s', posStr);
-
-      } else if (command.startsWith('draw')) {
-        await drawHandler(command, arg1, arg2);
-        ws.send(`${data}`);
-
-      } else if (command === 'prnt_scrn') {
-        const base64stringPNG = await psHandler();
-        // console.log('🚀 img:', img);
-        ws.send(`prnt_scrn ${base64stringPNG}`);
-        console.log('WebSocket Server -> prnt_scrn <base64 string (png buf)>');
-        // ws.send(`ps`);
+        } else {
+          duplex.write(data);
+          console.log('WebSocket Server -> %s', data);
+        }
       }
+    });
+
+    duplex.on('end', () => {
+      duplex.end();
+    });
+
+
+    ws.on('close', () => {
+      ws.close();
     });
 
     ws.send('Connected');
   });
 
-  wsServer.on('close', () => {});
+  wsServer.on('close', () => {
+    console.log('The connection has been closed successfully.');
+  });
 };
